Split up additionalPropertiesPostProcessor and PatternPropertiesPostProcessor to support serialization without adding the corresponding accessor methods
