--- src/SchemaProcessor/PostProcessor/Templates/AdditionalProperties/AdditionalPropertiesSerializer.phptpl
@@ -1,18 +1,0 @@
-/**
- * serialize all additional properties with the merge value strategy so all values stored in $this->additionalProperties
- * will be added to the top level of the serialization result
- */
-protected function serializeAdditionalProperties(): SerializedValue
-{
-    {% if serializerClass %}
-        $serializedValues = [];
-        foreach ($this->additionalProperties as $key => $value) {
-            $serializedValues[$key] = \{{ serializerClass }}::{{ serializerMethod }}($value, {{ serializerOptions }});
-        }
-    {% endif %}
-
-    return new SerializedValue(
-        {% if serializerClass %}$serializedValues{% else %}$this->additionalProperties{% endif %},
-        SerializedValue::STRATEGY_MERGE_VALUE
-    );
-}
