+++ src/SchemaProcessor/PostProcessor/Templates/AdditionalProperties/GetAdditionalProperties.phptpl
@@ -1,0 +1,9 @@
+/**
+ * Get additional properties
+ *
+ * @return {% if validationProperty %}{% if viewHelper.getTypeHintAnnotation(validationProperty, true) %}{{ viewHelper.getTypeHintAnnotation(validationProperty, true) }}{% endif %}{% else %}array{% endif %}
+ */
+public function getAdditionalProperties(): array
+{
+    return $this->_additionalProperties;
+}
