+++ src/SchemaProcessor/PostProcessor/Templates/AdditionalProperties/GetAdditionalProperty.phptpl
@@ -7,5 +7,5 @@
  */
 public function getAdditionalProperty(string $property){% if validationProperty %}{% if validationProperty.getType(true) %}: {{ viewHelper.getType(validationProperty, true) }}{% endif %}{% endif %}
 {
-    return $this->additionalProperties[$property] ?? null;
+    return $this->_additionalProperties[$property] ?? null;
 }
