+++ src/SchemaProcessor/PostProcessor/Templates/AdditionalProperties/RemoveAdditionalProperty.phptpl
@@ -19,7 +19,7 @@ public function removeAdditionalProperty(string $property): bool
         }
     }
 
-    if (!array_key_exists($property, $this->additionalProperties)) {
+    if (!array_key_exists($property, $this->_additionalProperties)) {
         return false;
     }
 
@@ -40,7 +40,7 @@ public function removeAdditionalProperty(string $property): bool
     {% endif %}
 
     unset($this->_rawModelDataInput[$property]);
-    unset($this->additionalProperties[$property]);
+    unset($this->_additionalProperties[$property]);
 
     return true;
 }
